+++ extern/vcpkg-overlays/triplets/arm64-ios-unreal.cmake
@@ -7,6 +7,6 @@ set(VCPKG_CMAKE_SYSTEM_NAME iOS)
 set(VCPKG_OSX_DEPLOYMENT_TARGET 15)
 
 # From Unreal Build Tool:
-# https://github.com/EpicGames/UnrealEngine/blob/5.2.1-release/Engine/Source/Programs/UnrealBuildTool/Platform/IOS/IOSToolChain.cs
+# https://github.com/EpicGames/UnrealEngine/blob/5.3.2-release/Engine/Source/Programs/UnrealBuildTool/Platform/IOS/IOSToolChain.cs
 set(VCPKG_CXX_FLAGS "-fvisibility=hidden")
 set(VCPKG_C_FLAGS "${VCPKG_CXX_FLAGS}")